refactor : 도메인 검증 클래스 패키지 리팩터링
