+++ src/main/java/com/example/board/domain/hobby/entity/Hobby.java
@@ -1,20 +1,27 @@
 package com.example.board.domain.hobby.entity;
 
+import com.example.board.domain.hobby.validator.HobbyValidator;
 import com.example.board.global.entity.BaseEntity;
-import jakarta.persistence.*;
-import jakarta.validation.constraints.NotNull;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-
-import static com.example.board.global.validator.HobbyValidation.validateHobbyTypeNull;
 
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Table(name = "hobby")
 @Entity
 public class Hobby extends BaseEntity {
+
+    private static final HobbyValidator validator = new HobbyValidator();
 
     @Id
     @GeneratedValue
@@ -23,13 +30,13 @@ public class Hobby extends BaseEntity {
 
     @Enumerated(EnumType.STRING)
     @Column(name = "hobby_type")
-    @NotNull(message = "{exception.entity.hobby.type}")
     private HobbyType hobbyType;
 
 
     @Builder
     public Hobby(HobbyType hobbyType) {
-        validateHobbyTypeNull(hobbyType);
+        validator.validateHobbyTypeNull(hobbyType);
         this.hobbyType = hobbyType;
     }
+
 }
