+++ src/main/java/com/example/board/domain/post/entity/Post.java
@@ -1,36 +1,38 @@
 package com.example.board.domain.post.entity;
 
+import static com.example.board.domain.post.dto.PostDto.*;
+
+import com.example.board.domain.post.validator.PostValidator;
 import com.example.board.domain.user.entity.User;
 import com.example.board.global.entity.BaseEntity;
-import jakarta.persistence.*;
-import jakarta.validation.constraints.NotNull;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.hibernate.validator.constraints.Length;
-
-import static com.example.board.domain.post.dto.PostDto.UpdatePostRequest;
-import static com.example.board.global.validator.PostValidator.validateContent;
-import static com.example.board.global.validator.PostValidator.validateTitle;
 
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Entity
 public class Post extends BaseEntity {
 
+    private static final PostValidator validator = new PostValidator();
+
     @Id @GeneratedValue
     @Column(name = "post_id", updatable = false)
     private Long id;
 
     @Column(nullable = false)
-    @NotNull(message = "{exception.entity.post.title.null}")
-    @Length(min = 1, message = "{exception.entity.post.title.length}")
     private String title;
 
     @Column(nullable = false)
-    @NotNull(message = "{exception.entity.post.content.null}")
-    @Length(min = 1, message = "{exception.entity.post.content.length}")
     private String content;
 
     @ManyToOne(fetch = FetchType.LAZY)
@@ -39,16 +41,16 @@ public class Post extends BaseEntity {
 
     @Builder
     public Post(String title, String content, User user) {
-        validateTitle(title);
-        validateContent(content);
+        validator.validateTitle(title);
+        validator.validateContent(content);
         this.title = title;
         this.content = content;
         this.user = user;
     }
 
     public void updatePost(UpdatePostRequest updatePostRequest) {
-        validateTitle(title);
-        validateContent(content);
+        validator.validateTitle(title);
+        validator.validateContent(content);
         this.title = updatePostRequest.title();
         this.content = updatePostRequest.content();
     }
