+++ src/main/java/com/example/board/domain/userhobby/validator/UserHobbyValidator.java
@@ -1,16 +1,17 @@
-package com.example.board.global.validator;
+package com.example.board.domain.userhobby.validator;
+
+import org.springframework.util.Assert;
 
 import com.example.board.domain.hobby.entity.Hobby;
 import com.example.board.domain.user.entity.User;
-import org.springframework.util.Assert;
 
 public class UserHobbyValidator {
 
-    public static void validateUserNull(User user) {
+    public void validateUserNull(User user) {
         Assert.notNull(user, "{exception.entity.user.null}");
     }
 
-    public static void validateHobbyNull(Hobby hobby) {
+    public void validateHobbyNull(Hobby hobby) {
         Assert.notNull(hobby, "{exception.entity.hobby.null}");
     }
 }
